+++ meta/src/store/config.rs
@@ -78,10 +78,7 @@ mod test {
 id = 1
 host = "127.0.0.1"
 port = 8901
-
-snapshot_path = "/tmp/cnosdb/meta/snapshot"
-journal_path = "/tmp/cnosdb/meta/journal"
-snapshot_per_events = 500
+data_path = "/tmp/cnosdb/meta"
 
 [log]
 level = "warn"
