+++ replication/src/entry_store.rs
@@ -141,3 +141,38 @@ impl EntryStorage for HeedEntryStorage {
         Ok(())
     }
 }
+
+mod test {
+    use heed::byteorder::BigEndian;
+    use heed::types::*;
+    use heed::Database;
+
+    #[test]
+    #[ignore]
+    fn test_heed_range() {
+        type BEU64 = U64<BigEndian>;
+
+        let path = "/tmp/cnosdb/8201-entry";
+        let env = heed::EnvOpenOptions::new()
+            .map_size(1024 * 1024 * 1024)
+            .max_dbs(128)
+            .open(path)
+            .unwrap();
+
+        let db: Database<OwnedType<BEU64>, OwnedSlice<u8>> =
+            env.create_database(Some("entries")).unwrap();
+
+        let mut wtxn = env.write_txn().unwrap();
+        let range = ..BEU64::new(4);
+        db.delete_range(&mut wtxn, &range).unwrap();
+        wtxn.commit().unwrap();
+
+        let reader = env.read_txn().unwrap();
+        let range = BEU64::new(0)..BEU64::new(1000000);
+        let iter = db.range(&reader, &range).unwrap();
+        for pair in iter {
+            let (index, _) = pair.unwrap();
+            println!("--- {}", index);
+        }
+    }
+}
