+++ replication/src/state_store.rs
@@ -236,6 +236,15 @@ impl StateStorage {
 
         Ok(())
     }
+
+    pub fn debug(&self) {
+        let reader = self.reader_txn().unwrap();
+        let iter = self.db.iter(&reader).unwrap();
+        for pair in iter {
+            let (key, val) = pair.unwrap();
+            println!("{}: {}", key, String::from_utf8_lossy(&val));
+        }
+    }
 }
 
 #[cfg(test)]
@@ -243,11 +252,11 @@ mod test {
     use std::fs;
     use std::path::Path;
 
-    use heed::byteorder::BigEndian;
     use heed::types::*;
     use heed::Database;
 
     #[test]
+    #[ignore]
     fn test_heed() {
         let path = "/tmp/cnosdb/test_heed";
         fs::create_dir_all(Path::new(&path)).unwrap();
@@ -278,39 +287,5 @@ mod test {
         }
 
         fs::remove_dir_all(path).unwrap();
-    }
-
-    #[test]
-    #[ignore]
-    fn test_heed_range() {
-        type BEU64 = U64<BigEndian>;
-
-        let path = "/tmp/cnosdb/meta/1_entry";
-        let env = heed::EnvOpenOptions::new()
-            .map_size(1024 * 1024 * 1024)
-            .max_dbs(128)
-            .open(path)
-            .unwrap();
-
-        let db: Database<OwnedType<BEU64>, OwnedSlice<u8>> =
-            env.create_database(Some("ttttt")).unwrap();
-
-        let mut wtxn = env.write_txn().unwrap();
-        let range = ..BEU64::new(80);
-        db.delete_range(&mut wtxn, &range).unwrap();
-        wtxn.commit().unwrap();
-
-        let mut wtxn = env.write_txn().unwrap();
-        let range = BEU64::new(110)..;
-        db.delete_range(&mut wtxn, &range).unwrap();
-        wtxn.commit().unwrap();
-
-        let reader = env.read_txn().unwrap();
-        let range = BEU64::new(0)..BEU64::new(1000000);
-        let iter = db.range(&reader, &range).unwrap();
-        for pair in iter {
-            let (index, _) = pair.unwrap();
-            println!("--- {}", index);
-        }
     }
 }
